Started working on updating the quasar model.
The biggest thing is that following a merger event, the central galaxy
BH accretes cold gas instantly.  While this is ok for something like
Millennium, I'm worried that this won't be applicable for Kali.

To this end, in this new model I follow something similar to Qin 2017's
implentation in MERAXES.  I calculate the amount of mass that should be
accreted by the black hole then compare that to the Eddington Mass.  If
the Eddington Mass is greater than the accreted mass then all the mass
will be accreted in a timestep otherwise only mass equal to the
Eddington Mass will be accreted.  Regardless of this, I spread this mass
accreted over the substeps.

In addition to altering the accretion rate to be smooth, I also include
the fractional quasar ejection.  During each accretion event I say that
a quasar pulses with energy from the amount of mass accreted.  Then I
compare this amount of energy with the thermal energy of the reservoirs
and eject/reheat mass proportional to the RATIO of the quasar energy to
the thermal energies.